requirements.txt
Flask
openai
python-dotenv
requests
gunicorn